fixed typo

+++ service-worker.js
@@ -13,7 +13,7 @@ const CACHE_NAME = APP_PREFIX + VERSION;
 self.addEventListener('install', function(e) {
     e.waitUntil(
         caches.open(CACHE_NAME).then(function(cache) {
-            console.logf('installing cache : ' + CACHE_NAME);
+            console.log('installing cache : ' + CACHE_NAME);
             return cache.addAll(FILES_TO_CACHE)
         })
     )
